+++ package.json
@@ -38,6 +38,7 @@
     "react-icons": "^3.11.0",
     "react-modal": "^3.11.2",
     "react-progressbar": "^15.4.1",
+    "react-router-dom": "^5.2.0",
     "react-star-ratings": "^2.3.0",
     "sequelize": "^6.3.5",
     "underscore": "^1.11.0"
